Avoid false negative in test
Conversion to primitive int was removed while
migrating to JUnit 5, and this commit restores it

+++ jgrapht-io/src/test/java/org/jgrapht/nio/graphml/SimpleGraphMLEdgeListImporterTest.java
@@ -133,7 +133,7 @@ public class SimpleGraphMLEdgeListImporterTest
             assertEquals(edge[0], e.getFirst());
             assertEquals(edge[1], e.getSecond());
             if (i < 2) {
-                assertEquals(edge[2], collected.get(i).getThird());
+                assertEquals(edge[2].intValue(), collected.get(i).getThird());
             } else {
                 assertNull(e.getThird());
